Strings - add formatted string
add formatted string that returns a string based on a format string and
replacing certain tokens with the arguments passed in a vector. this
function is templated so that any argument type can be used. every
argument is passed through stringFrom()
fixed trimWhitespace*() functions as they should not require
charactersToTrim since they use the presetWhitespaceCharacters