add operators to Rgb class
added mathematical operators to pl::Color::Rgb class which allows
adding, subtracting and multiplication of other Rgb colours as well as
multiplying and dividing by a scalar value.

added a clampStandardRange function to Rgb that will clamp the r, g and
b values to be clamped to the standard range (0.0 to 1.0)

altered TweenPiecewise so that it converts position to a float instead
of a double when using linear interpolation so that objects that can
only be multiplied by floats (but not doubles) can be interpolated, such
as SFML vectors.