add Ranges pull
(re-)add pulling to the range types. originally, pulling was a free
function that took the hook and range as parameters whereas now it is a
method of the range class that takes the hook as a parameter.
pulling has been added to both the Range and RangeArea classes where
RangeArea can do three types of pulling: normal (RangeArea), horizontal
and vertical (Range).
Pulling a different type (horizontal or vertical Range in RangeArea)
returns a copy of that range whereas pulling the same type returns a
reference to the actual object.