update StringFrom and Sfml StringFrom
*StringFrom*
added ability to add an extra parameter to stringFrom to specify the
number of decimal places to show (uses string stream internally). this
function is a template so can take any type; it will be casted to a long
double.

*Sfml/StringFrom*
added function to return a string from an sf::Time directly. returns the
value in seconds as well as the string " seconds" i.e. ("x seconds")