Ranges - added contains own range type
Added ability to test if a range is contained within another range.
i.e. a Range can test if it contains another Range and a RangeArea can
test if contains another RangeArea.
Uses RangeBoundaries/RangeAreaBoundaries in a similar way to contains()
for a single value.
The ranges' types must be exactly equal.