add TextureAtlas
added texture atlas class that can store information about which part of texture to use (Frame).

stored are:
rect - the texture rectangle; the main use of a texture atlas
origin - ability to store a specific point for each frame
texture index - ability to store an index referencing which texture should be used for this frame (allows TextureAtlas to reference over multiple textures)
category - ability to categorize/group frames
rotate/flip - bitwise enum class allowing any manipulation of the rectangle (rotation of 90 degrees or any flips)

note that the atlas stores the information but does not apply the information.

also available is the ability to get a vector of pointers to a selection of the frames:
all - every frame
all frames in category - every frame that has a category that matches the given category
all frames with texture index - every frame that has a texture index that matches the given one
all frames in category with texture index - every frame that matches both the given category and the given texture index