add SFML Image Resize processing
added ability to resize an sf::Image (or a given part/crop of it) to a given size. It returns the resulting resized sf::Image without affecting the original image.

multiple choices are available for the method of resizing including:
Draw - draws to a render texture and then loads the image from the render texture.
Pixel - uses only the exact pixels from the original image. pixels are dropped when shrinking and duplicated when expanding. this is useful for duplication
Nearest Neighbour - similar to Pixel but uses the pixel nearest the target. this means it is slightly offset from pixel but represents the original image's position more accurately
Bilinear - bilinear interpolation that can create more data than was originally available to reduce jagged edges

For Draw, there are two options that can be used separately, combined or ignored: Smooth and AA. Smooth uses a smoothed texture to draw onto the render texture. AA activates the anti-alias of the render texture. So, Smooth + AA uses a smoothed texture to draw onto an anti-aliased render texture.

Draw_Smooth gives results similar to Bilinear.

Note that Draw_Smooth requires the ability to create a render texture of the required destination size whereas Bilinear does not. However, Bilinear can be much slower as it's calculated manually on the CPU whereas Draw_Smooth uses the GPU.